Conflicts:
	package.json

+++ tasks/bust_requirejs_cache.js
@@ -22,6 +22,12 @@ var ignoreMatch = function(src, patterns) {
 	return false;
 }
 
+var getHash = function(fileContent){
+	var shasum = crypto.createHash('md5');
+	var hash = shasum.update(fileContent).digest('hex');
+	return hash;
+}
+
 module.exports = function(grunt) {
 
 	// Please see the Grunt documentation for more information regarding task
@@ -39,7 +45,7 @@ module.exports = function(grunt) {
 
 		var options = this.options();
 		var srcUrl = options.appDir + "/" + baseUrl;
-		grunt.log.writeln("src url: " + srcUrl);
+		grunt.log.debug("src url: " + srcUrl);
 
 
 		//grunt.log.writeln(util.inspect(options));
@@ -71,20 +77,29 @@ module.exports = function(grunt) {
 					var exp = /(['"])([\w-\/]+)(['"])/ig;
 					var result = value.replace(exp, function($0, $1, moduleName, $3) {
 						if (!/\.js$/.test(moduleName) && ! ignoreMatch(moduleName, options.ignorePatterns)) {
-							var filepath = path.join(srcUrl, moduleName) + '.js';
-							//grunt.log.writeln(filepath);
-							var shasum = crypto.createHash('md5');
+							//如果已经加过指纹，直接返回
+							if(resourceMap[moduleName]){
+								return $1 + resourceMap[moduleName] + $3;
+							}
 
+							var filepath = path.join(srcUrl, moduleName) + '.js';
+							grunt.log.debug(filepath);
+							
+
+							if(!grunt.file.exists(filepath)){
+								grunt.log.warn('target file "' + filepath + '" not found');
+								return $1 + moduleName + $3;
+							}
 							var fileContent = grunt.file.read(filepath);
 
-							var hash = shasum.update(fileContent).digest('hex');
+							var hash = getHash(fileContent);
 
-							var newFileContent = fileContent.replace(moduleName, moduleName + "." + hash);
-							grunt.file.write(filepath, newFileContent);
+							// var newFileContent = fileContent.replace(moduleName, moduleName + "." + hash);
+							// grunt.file.write(filepath, fileContent);
 
 							var moduleName = resourceMap[moduleName] = moduleName + '.' + hash;
 
-							grunt.log.writeln(filepath);
+							grunt.log.debug(filepath);
 
 							fs.renameSync(filepath, filepath.replace('.js', '.' + hash + '.js'));
 							//grunt.log.writeln(hash);
@@ -96,24 +111,54 @@ module.exports = function(grunt) {
 					return result;
 				});
 
-				grunt.log.writeln(hashedMatches);
-				requireMatches.forEach(function(value, index){
-					if(value != hashedMatches[index]){
-						src = src.replace(value, hashedMatches[index]);
-					}
-				});
+				grunt.log.debug(hashedMatches);
+				// requireMatches.forEach(function(value, index){
+				// 	if(value != hashedMatches[index]){
+				// 		src = src.replace(value, hashedMatches[index]);
+				// 	}
+				// });
 
 			}
 
 
 			// Write the destination file.
-			grunt.file.write(f.dest, src);
+			// grunt.file.write(f.dest, src);
 
 			// Print a success message.
 			grunt.log.writeln('File "' + f.dest + '" created.');
 		});
-		grunt.log.writeln(JSON.stringify(resourceMap));
+
+		var resourceMapFileContent = JSON.stringify(resourceMap);
+		grunt.log.debug(resourceMapFileContent);
+
+		//获取requirejs cache bust生成的js模块的source map
+		var rsconfigPath = this.options().dist + '/js/rs-config.js';
+        // var resourceMap = grunt.file.read(this.options().dist + '/resource-map.json');
+        var modulePath = 'requirejs.config({"paths": ' + resourceMapFileContent + "});";
+        var rsConfig = grunt.file.read(rsconfigPath);
+
+        grunt.file.write(rsconfigPath, rsConfig + modulePath);
+        grunt.log.writeln("update rs-config.js");
+        var rsconfigHash = getHash(rsConfig + modulePath);
+        fs.renameSync(rsconfigPath, rsconfigPath.replace('.js', '.' + rsconfigHash + '.js'));
+
+        resourceMap['rs-config'] = 'rs-config.' + rsconfigHash;
 		grunt.file.write(options.appDir + '/resource-map.json', JSON.stringify(resourceMap));
+
+        this.files.forEach(function(f){
+        	var src = '';
+
+			if (!grunt.file.exists(f.src[0])) {
+				grunt.log.warn('Source file "' + f.src[0] + '" not found.');
+			} else {
+				src = grunt.file.read(f.src[0]);
+			}
+
+			var src = src.replace(/rs-config/ig, 'rs-config.' + rsconfigHash);
+
+			grunt.file.write(f.dest, src);
+
+        });
 	});
 
 };
